move decorators to decorators submodule, add log decorator
