+++ bot.py
@@ -31,19 +31,7 @@ dispatcher = updater.dispatcher
 
 
 
-def log(func):
-    '''decorator that logs who said what to the bot'''
-    @wraps(func)
-    def wrapped(update, context, *args, **kwargs):
-        id = update.effective_user.id
-        name = update.effective_user.username
-        context.user_data['meta'] = {
-            'last_talked': update.effective_message['date'],
-            'user_details': update.effective_message.to_dict()['from']
-        }
-        logging.info(f'{name} ({id}) said:\n{update.effective_message.text}')
-        return func(update, context, *args, **kwargs)
-    return wrapped
+
 
 
 def restricted(func):
