+++ commands/__init__.py
@@ -1,2 +1,1 @@
 from .error import error
-from .send_typing_action import send_typing_action
